+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [1.0.0(7)] - 2018-07-09
 ### Added
 - User agent in Open Radar API request #4
 - Open Radar user API key #5